restructure player
