+++ src/common/interaction/util.js
@@ -1,4 +1,4 @@
-import * as YoutubeApi from './youtube_api';
+import * as YoutubeApi from 'core/youtube_api';
 
 export const videosRate = (videoId, rating) => {
   let params = {
