+++ src/modules/player/components/new_comment.jsx
@@ -1,9 +1,8 @@
-import React             from 'react';
-import PropTypes         from 'prop-types';
-import YT_API_KEY        from '../../../config/api_key';
+import React from 'react';
+import PropTypes from 'prop-types';
+import YT_API_KEY from '../../../config/api_key';
 
 class NewComment extends React.Component {
-
   constructor(props) {
     super(props);
 
@@ -22,7 +21,6 @@ class NewComment extends React.Component {
 
   handleSubmit(e) {
     e.preventDefault();
-
   }
 
   render() {
