+++ src/modules/player/components/related_list_item.jsx
@@ -1,5 +1,5 @@
-import React                 from 'react';
-import PropTypes             from 'prop-types';
+import React from 'react';
+import PropTypes from 'prop-types';
 import { hashHistory, Link } from 'react-router';
 
 class RelatedListItem extends React.Component {
