+++ src/modules/player/util.js
@@ -1,4 +1,4 @@
-import * as YoutubeApi from '../youtube_api';
+import * as YoutubeApi from 'core/youtube_api';
 import YT_API_KEY from '../../../config/api_key';
 import { createUrlParams } from 'helpers';
 
@@ -30,16 +30,6 @@ export const fetchVideoRating = videoId => {
   };
   return YoutubeApi.videosGetRating(params);
 };
-// 
-// export const fetchVideos = (query, nextPageToken = null) => {
-//   let params = {
-//     q: query,
-//     type: 'video',
-//     pageToken: nextPageToken,
-//   };
-//
-//   return YoutubeApi.search(params);
-// };
 
 export const fetchRelated = videoId => {
   let params = {
@@ -49,12 +39,3 @@ export const fetchRelated = videoId => {
   };
   return YoutubeApi.search(params);
 };
-//
-// export const fetchVideoStats = videos => {
-//   let params = {
-//     part: 'statistics',
-//     id: videos.items.map(item => item.id.videoId).join(','),
-//   };
-//
-//   return YoutubeApi.videos(params);
-// };
